requirements.txt
asgiref==3.8.1
dependency-injector==4.46.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
Markdown==3.8
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
